fix: avoid returning false by default when executing the scanCard operation
When the scanCard method is executed without the permission for the camera usage, a popup is showed to the user in order to request the required permission. However, the plugin was returning false as the result of the method execution. This behaviour was invoking the fail callback that was passed and, at the same time, if the user gave the required permission, it was opening the scanning view, which allowed the user to scan a card and fulfil the required data, but no data was being passed to the success callback.

References: https://outsystemsrd.atlassian.net/browse/RNMT-3665